--- src/main/java/Deadline.java
@@ -1,21 +1,0 @@
-public class Deadline extends Task {
-
-    import main.java.Task;
-
-    private String time;
-
-        public Deadline(String task, String time) {
-            super(task);
-            this.time = time;
-        }
-
-        @Override
-        public String toString() {
-            return "[D]" +
-                    (this.checkStatus() ? "[X] " : "[ ] ") +
-                    super.toString() +
-                    " (by: " +
-                    this.time +
-                    ")";
-        }
-}
